% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housingdecay-app.R
\docType{data}
\name{allresponses}
\alias{allresponses}
\title{allresponses}
\format{
A data frame with 6 questions to be joined with the survey respondents
}
\source{
Own research
}
\usage{
allresponses
}
\description{
Dummy data used to reformat the matrix question in long format
}
\keyword{datasets}
