% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housingdecay-app.R
\docType{data}
\name{ds}
\alias{ds}
\title{DS}
\format{
A data frame with 164 images and rows to be filled via Shiny by the survey respondents
}
\source{
Google Maps
}
\usage{
ds
}
\description{
Dataset with images from Google Maps
}
\keyword{datasets}
