% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housingdecay-app.R
\docType{data}
\name{matrix_question}
\alias{matrix_question}
\title{matrix_question}
\format{
A data frame with 1 question to be completed via Shiny by the survey respondents
}
\source{
Own research
}
\usage{
matrix_question
}
\description{
Question in survey format
}
\keyword{datasets}
