% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housingdecay-app.R
\docType{data}
\name{radioB_question}
\alias{radioB_question}
\title{radioB_question}
\format{
A data frame with 5 questions to be completed via Shiny by the survey respondents
}
\source{
Own research
}
\usage{
radioB_question
}
\description{
Questions in survey format
}
\keyword{datasets}
